feat: split broker.py into the broker/ domain package
