--- cafleet/src/cafleet/broker.py
@@ -1,1246 +1,0 @@
-"""Sync SQLAlchemy data-access layer shared by the CLI and WebUI."""
-
-import json
-from datetime import UTC, datetime
-
-import click
-from sqlalchemy import and_, delete, exists, func, or_, select, update
-from sqlalchemy.dialects.sqlite import insert as sqlite_insert
-
-from cafleet.config import settings
-from cafleet.db.engine import get_sync_sessionmaker
-from cafleet.db.models import Agent, AgentPlacement, Fleet, Task
-from cafleet.multiplexer import MultiplexerContext
-
-_DIRECTOR_NAME = "Director"
-_DIRECTOR_DESCRIPTION = "Root Director for this fleet"
-
-ADMINISTRATOR_KIND = "builtin-administrator"
-
-_TASK_COLUMNS = tuple(Task.__table__.columns.keys())
-
-_NOT_BROADCAST_SUMMARY = Task.type != "broadcast_summary"
-
-
-def _is_administrator(agent_card_json: str | None) -> bool:
-    if not agent_card_json:
-        return False
-    try:
-        kind = json.loads(agent_card_json).get("cafleet", {}).get("kind")
-    except ValueError:
-        return False
-    return kind == ADMINISTRATOR_KIND
-
-
-def _now_iso() -> str:
-    return datetime.now(UTC).isoformat()
-
-
-def _placement_dict(row) -> dict:
-    return {
-        "director_agent_id": row.director_agent_id,
-        "tmux_session": row.tmux_session,
-        "tmux_window_id": row.tmux_window_id,
-        "tmux_pane_id": row.tmux_pane_id,
-        "coding_agent": row.coding_agent,
-        "created_at": row.created_at,
-    }
-
-
-def _agent_is_active_in_fleet(session, agent_id: int, fleet_id: int) -> bool:
-    return session.execute(
-        select(
-            exists().where(
-                Agent.agent_id == agent_id,
-                Agent.fleet_id == fleet_id,
-                Agent.status == "active",
-            )
-        )
-    ).scalar_one()
-
-
-def _try_notify_recipient(
-    session, *, recipient_id: int, sender_id: int, task_dict: dict
-) -> bool:
-    """Best-effort inline-preview keystroke for the recipient's pane.
-
-    Keystrokes a 2-line preview of the message itself into the recipient's
-    pane — the recipient's TUI processes the keystrokes as a fresh user-turn
-    input and the recipient acks via
-    ``cafleet message ack --task-id <id>``. The queue remains the source
-    of truth; failures are swallowed.
-    """
-    if recipient_id == sender_id:
-        return False
-    pane_id = session.execute(
-        select(AgentPlacement.tmux_pane_id).where(
-            AgentPlacement.agent_id == recipient_id
-        )
-    ).scalar_one_or_none()
-    if pane_id is None:
-        return False
-    # Local import so tests that monkeypatch
-    # ``cafleet.multiplexer.tmux.TmuxMultiplexer.send_inline_preview``
-    # get picked up on every call rather than bound once at broker import.
-    from cafleet.multiplexer.tmux import TmuxMultiplexer
-
-    # Truncate before keystroking so a multi-KB body cannot dump itself into
-    # the recipient's pane. Mirrors output.truncate_text's contract: same
-    # limit (``settings.max_text_len`` / ``CAFLEET_MAX_TEXT_LEN``, default
-    # 200) and same single-codepoint U+2026 suffix on overflow.
-    preview_text = task_dict["text"]
-    if len(preview_text) > settings.max_text_len:
-        preview_text = preview_text[: settings.max_text_len] + "…"
-
-    return TmuxMultiplexer().send_inline_preview(
-        target_pane_id=pane_id,
-        task_id=task_dict["task_id"],
-        sender_id=sender_id,
-        ts=task_dict["status_timestamp"],
-        text=preview_text,
-    )
-
-
-def create_fleet(
-    label: str | None = None,
-    *,
-    director_context: MultiplexerContext,
-    coding_agent: str,
-) -> dict:
-    """Atomically bootstrap a fleet with its root Director and Administrator.
-
-    The fleet row is written first with ``director_agent_id=NULL`` and
-    back-filled once the Director's agent row exists, so the column is
-    DB-nullable even though the post-bootstrap invariant is NOT NULL.
-
-    Args:
-        label: Optional human-readable label for the fleet.
-        director_context: Resolved tmux pane identity for the root Director,
-            obtained via ``Multiplexer.context_discovery``.
-        coding_agent: Operator-declared metadata that lands in the root
-            Director's ``placement.coding_agent`` column. The CLI is the only
-            caller and always supplies it (default ``'claude'`` lives at the
-            Click layer).
-
-    Returns:
-        A dict carrying ``fleet_id``, ``label``, ``created_at``,
-        ``administrator_agent_id``, and a ``director`` sub-dict with the
-        Director's identity and placement metadata.
-    """
-    created_at = _now_iso()
-    director_card = {
-        "name": _DIRECTOR_NAME,
-        "description": _DIRECTOR_DESCRIPTION,
-        "skills": [],
-    }
-    director_placement = {
-        "director_agent_id": None,
-        "tmux_session": director_context.session,
-        "tmux_window_id": director_context.window_id,
-        "tmux_pane_id": director_context.pane_id,
-        "coding_agent": coding_agent,
-        "created_at": created_at,
-    }
-
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        fleet = Fleet(
-            label=label,
-            created_at=created_at,
-            deleted_at=None,
-            director_agent_id=None,
-        )
-        session.add(fleet)
-        session.flush()
-        fleet_id = fleet.fleet_id
-        director = Agent(
-            fleet_id=fleet_id,
-            name=_DIRECTOR_NAME,
-            description=_DIRECTOR_DESCRIPTION,
-            status="active",
-            registered_at=created_at,
-            deregistered_at=None,
-            agent_card_json=json.dumps(director_card),
-        )
-        session.add(director)
-        session.flush()
-        director_agent_id = director.agent_id
-        session.add(AgentPlacement(agent_id=director_agent_id, **director_placement))
-        session.flush()
-        session.execute(
-            update(Fleet)
-            .where(Fleet.fleet_id == fleet_id)
-            .values(director_agent_id=director_agent_id)
-        )
-        administrator_card = {
-            "name": "Administrator",
-            "description": f"Built-in administrator agent for fleet {fleet_id}",
-            "skills": [],
-            "cafleet": {"kind": ADMINISTRATOR_KIND},
-        }
-        administrator = Agent(
-            fleet_id=fleet_id,
-            name=administrator_card["name"],
-            description=administrator_card["description"],
-            status="active",
-            registered_at=created_at,
-            deregistered_at=None,
-            agent_card_json=json.dumps(administrator_card),
-        )
-        session.add(administrator)
-        session.flush()
-        administrator_agent_id = administrator.agent_id
-
-    return {
-        "fleet_id": fleet_id,
-        "label": label,
-        "created_at": created_at,
-        "administrator_agent_id": administrator_agent_id,
-        "director": {
-            "agent_id": director_agent_id,
-            "name": _DIRECTOR_NAME,
-            "description": _DIRECTOR_DESCRIPTION,
-            "registered_at": created_at,
-            "placement": director_placement,
-        },
-    }
-
-
-def list_fleets() -> list[dict]:
-    """Return non-soft-deleted fleets with their active agent counts."""
-    stmt = (
-        select(
-            Fleet.fleet_id,
-            Fleet.director_agent_id,
-            Fleet.label,
-            Fleet.created_at,
-            func.count(Agent.agent_id).label("agent_count"),
-        )
-        .select_from(Fleet)
-        .outerjoin(
-            Agent,
-            and_(
-                Agent.fleet_id == Fleet.fleet_id,
-                Agent.status == "active",
-            ),
-        )
-        .where(Fleet.deleted_at.is_(None))
-        .group_by(
-            Fleet.fleet_id,
-            Fleet.director_agent_id,
-            Fleet.label,
-            Fleet.created_at,
-        )
-        .order_by(Fleet.created_at.desc(), Fleet.fleet_id.asc())
-    )
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-    return [
-        {
-            "fleet_id": row.fleet_id,
-            "director_agent_id": row.director_agent_id,
-            "label": row.label,
-            "created_at": row.created_at,
-            "agent_count": row.agent_count,
-        }
-        for row in rows
-    ]
-
-
-def get_fleet(fleet_id: int) -> dict | None:
-    """Return the fleet row (including soft-deleted) or None.
-
-    The returned dict exposes ``deleted_at`` so callers can distinguish a
-    missing fleet from a soft-deleted one — ``register_agent`` relies on
-    this to reject soft-deleted fleets with a different error message.
-
-    Args:
-        fleet_id: Fleet id to look up.
-
-    Returns:
-        Dict with ``fleet_id``, ``label``, ``created_at``, ``deleted_at``,
-        and ``director_agent_id``, or ``None`` if no row exists.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        result = session.execute(select(Fleet).where(Fleet.fleet_id == fleet_id))
-        row = result.scalar_one_or_none()
-    if row is None:
-        return None
-    return {
-        "fleet_id": row.fleet_id,
-        "label": row.label,
-        "created_at": row.created_at,
-        "deleted_at": row.deleted_at,
-        "director_agent_id": row.director_agent_id,
-    }
-
-
-def delete_fleet(fleet_id: int) -> dict:
-    """Soft-delete a fleet and deregister its agents, in one transaction.
-
-    Tasks are left untouched so audit history survives. Idempotent: re-running
-    against an already-deleted row short-circuits on the ``deleted_at IS NULL``
-    guard and returns ``deregistered_count=0``.
-
-    Args:
-        fleet_id: Fleet id to soft-delete.
-
-    Returns:
-        Dict with ``deregistered_count`` — the number of agents flipped from
-        ``active`` to ``deregistered`` by this call.
-
-    Raises:
-        click.ClickException: If the fleet does not exist.
-    """
-    now = _now_iso()
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        fleet_exists = session.execute(
-            select(exists().where(Fleet.fleet_id == fleet_id))
-        ).scalar_one()
-        if not fleet_exists:
-            # ClickException exits 1 (matching ``fleet show``); UsageError
-            # would print a Usage: banner + exit 2, wrong for a runtime miss.
-            raise click.ClickException(f"fleet '{fleet_id}' not found.")
-
-        soft_deleted = session.execute(
-            update(Fleet)
-            .where(
-                Fleet.fleet_id == fleet_id,
-                Fleet.deleted_at.is_(None),
-            )
-            .values(deleted_at=now)
-            .returning(Fleet.fleet_id)
-        ).all()
-        if not soft_deleted:
-            return {"deregistered_count": 0}
-
-        deregistered = session.execute(
-            update(Agent)
-            .where(
-                Agent.fleet_id == fleet_id,
-                Agent.status == "active",
-            )
-            .values(status="deregistered", deregistered_at=now)
-            .returning(Agent.agent_id)
-        ).all()
-        deregistered_count = len(deregistered)
-        agents_in_fleet = select(Agent.agent_id).where(Agent.fleet_id == fleet_id)
-        session.execute(
-            delete(AgentPlacement).where(AgentPlacement.agent_id.in_(agents_in_fleet))
-        )
-
-    return {"deregistered_count": deregistered_count}
-
-
-def register_agent(
-    fleet_id: int,
-    name: str,
-    description: str,
-    skills: list[dict] | None = None,
-    placement: dict | None = None,
-) -> dict:
-    """Register a new agent in the fleet and optionally create its placement.
-
-    Rejects soft-deleted fleets with a message that differs from the
-    "not found" case so callers can surface the right recovery hint. When
-    ``placement`` is supplied, the named Director must be active in the
-    same fleet, must not be the Administrator, and must be the fleet's root
-    Director (``fleets.director_agent_id``). A fleet has exactly one Director:
-    nested teams are forbidden, so every member placement's
-    ``director_agent_id`` equals the fleet root and a non-root value is
-    rejected.
-
-    Args:
-        fleet_id: Fleet id the new agent belongs to.
-        name: Short human-identifiable label.
-        description: One-sentence purpose statement.
-        skills: Optional list of skill dicts persisted into the agent's
-            ``agent_card_json`` blob.
-        placement: Optional dict carrying ``director_agent_id``,
-            ``tmux_session``, ``tmux_window_id``, ``tmux_pane_id``, and
-            ``coding_agent``. When present, an ``AgentPlacement`` row is
-            created alongside the agent.
-
-    Returns:
-        Dict with ``agent_id``, ``name``, and ``registered_at``.
-
-    Raises:
-        click.UsageError: If the fleet does not exist, is soft-deleted, the
-            named Director is not active in the same fleet, or the placement
-            ``director_agent_id`` is not the fleet's root Director.
-        click.ClickException: If the named Director is the built-in
-            Administrator.
-    """
-    sess = get_fleet(fleet_id)
-    if sess is None:
-        raise click.UsageError(f"Fleet '{fleet_id}' not found.")
-    if sess["deleted_at"] is not None:
-        raise click.UsageError(f"fleet {fleet_id} is deleted")
-
-    registered_at = _now_iso()
-    agent_card = {
-        "name": name,
-        "description": description,
-        "skills": skills or [],
-    }
-
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        if placement is not None:
-            director_id = placement["director_agent_id"]
-            director_card = session.execute(
-                select(Agent.agent_card_json).where(
-                    Agent.agent_id == director_id,
-                    Agent.fleet_id == fleet_id,
-                    Agent.status == "active",
-                )
-            ).scalar_one_or_none()
-            if director_card is None:
-                raise click.UsageError(
-                    f"Director agent '{director_id}' not found or not active "
-                    f"in fleet '{fleet_id}'."
-                )
-            if _is_administrator(director_card):
-                raise click.ClickException("Administrator cannot be a director")
-            root_director_id = sess["director_agent_id"]
-            if director_id != root_director_id:
-                raise click.UsageError(
-                    f"nested teams are not supported; placement director_agent_id "
-                    f"{director_id} must equal the fleet root "
-                    f"Director {root_director_id}."
-                )
-
-        agent = Agent(
-            fleet_id=fleet_id,
-            name=name,
-            description=description,
-            status="active",
-            registered_at=registered_at,
-            agent_card_json=json.dumps(agent_card),
-        )
-        session.add(agent)
-        session.flush()
-        agent_id = agent.agent_id
-        if placement is not None:
-            session.add(
-                AgentPlacement(
-                    agent_id=agent_id,
-                    director_agent_id=placement["director_agent_id"],
-                    tmux_session=placement["tmux_session"],
-                    tmux_window_id=placement["tmux_window_id"],
-                    tmux_pane_id=placement["tmux_pane_id"],
-                    coding_agent=placement["coding_agent"],
-                    created_at=registered_at,
-                )
-            )
-
-    return {
-        "agent_id": agent_id,
-        "name": name,
-        "registered_at": registered_at,
-    }
-
-
-def get_agent(agent_id: int, fleet_id: int) -> dict | None:
-    """Return the active agent's detail (with placement) or None.
-
-    Args:
-        agent_id: Agent id to look up.
-        fleet_id: Fleet id the agent must belong to.
-
-    Returns:
-        Dict with ``agent_id``, ``name``, ``description``, ``status``,
-        ``registered_at``, ``kind`` (``"user"`` or the Administrator kind),
-        and ``placement`` (the placement sub-dict or ``None``). Returns
-        ``None`` if no active agent matches.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        agent = session.execute(
-            select(Agent).where(
-                Agent.agent_id == agent_id,
-                Agent.fleet_id == fleet_id,
-                Agent.status == "active",
-            )
-        ).scalar_one_or_none()
-
-        if agent is None:
-            return None
-
-        placement_row = session.execute(
-            select(AgentPlacement).where(AgentPlacement.agent_id == agent_id)
-        ).scalar_one_or_none()
-
-    result: dict = {
-        "agent_id": agent.agent_id,
-        "name": agent.name,
-        "description": agent.description,
-        "status": agent.status,
-        "registered_at": agent.registered_at,
-        "kind": (
-            ADMINISTRATOR_KIND if _is_administrator(agent.agent_card_json) else "user"
-        ),
-        "placement": None,
-    }
-    if placement_row is not None:
-        result["placement"] = _placement_dict(placement_row)
-    return result
-
-
-def list_agents(fleet_id: int) -> list[dict]:
-    """Return all active agents in the fleet."""
-    stmt = select(
-        Agent.agent_id,
-        Agent.name,
-        Agent.description,
-        Agent.registered_at,
-    ).where(
-        Agent.fleet_id == fleet_id,
-        Agent.status == "active",
-    )
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-    return [
-        {
-            "agent_id": row.agent_id,
-            "name": row.name,
-            "description": row.description,
-            "status": "active",
-            "registered_at": row.registered_at,
-        }
-        for row in rows
-    ]
-
-
-def deregister_agent(agent_id: int) -> bool:
-    """Soft-delete the agent and drop its placement.
-
-    Args:
-        agent_id: Agent id to deregister.
-
-    Returns:
-        ``True`` if a row was flipped from ``active`` to ``deregistered``;
-        ``False`` if no matching active agent existed.
-
-    Raises:
-        click.UsageError: If ``agent_id`` is the root Director of any
-            fleet — torn down via ``cafleet fleet delete`` instead.
-        click.ClickException: If ``agent_id`` is the built-in Administrator.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        is_root_director = session.execute(
-            select(exists().where(Fleet.director_agent_id == agent_id))
-        ).scalar_one()
-        if is_root_director:
-            raise click.UsageError(
-                "cannot deregister the root Director; "
-                "use 'cafleet fleet delete' instead"
-            )
-
-        card_json = session.execute(
-            select(Agent.agent_card_json).where(Agent.agent_id == agent_id)
-        ).scalar_one_or_none()
-        if card_json is not None and _is_administrator(card_json):
-            raise click.ClickException("Administrator cannot be deregistered")
-        deregistered = session.execute(
-            update(Agent)
-            .where(
-                Agent.agent_id == agent_id,
-                Agent.status == "active",
-            )
-            .values(
-                status="deregistered",
-                deregistered_at=_now_iso(),
-            )
-            .returning(Agent.agent_id)
-        ).all()
-        if deregistered:
-            session.execute(
-                delete(AgentPlacement).where(AgentPlacement.agent_id == agent_id)
-            )
-    return bool(deregistered)
-
-
-def update_placement_pane_id(agent_id: int, pane_id: str) -> dict | None:
-    """Patch the agent's placement with a freshly resolved tmux pane id.
-
-    Called after ``split_window`` returns the spawned pane's id so the
-    placement row reflects the live pane rather than the placeholder used
-    during the initial INSERT.
-
-    Args:
-        agent_id: Agent id whose placement should be updated.
-        pane_id: New ``tmux_pane_id`` value.
-
-    Returns:
-        The refreshed placement dict, or ``None`` if no placement row was
-        affected.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        updated = session.execute(
-            update(AgentPlacement)
-            .where(AgentPlacement.agent_id == agent_id)
-            .values(tmux_pane_id=pane_id)
-            .returning(AgentPlacement.agent_id)
-        ).first()
-        if updated is None:
-            return None
-        row = session.execute(
-            select(AgentPlacement).where(AgentPlacement.agent_id == agent_id)
-        ).scalar_one()
-    return _placement_dict(row)
-
-
-def _base_members_select(fleet_id: int):
-    return (
-        select(
-            Agent.agent_id,
-            Agent.name,
-            Agent.description,
-            Agent.status,
-            Agent.registered_at,
-            AgentPlacement.director_agent_id,
-            AgentPlacement.tmux_session,
-            AgentPlacement.tmux_window_id,
-            AgentPlacement.tmux_pane_id,
-            AgentPlacement.coding_agent,
-            AgentPlacement.created_at,
-        )
-        .join(AgentPlacement, Agent.agent_id == AgentPlacement.agent_id)
-        .where(
-            Agent.fleet_id == fleet_id,
-            Agent.status == "active",
-            AgentPlacement.director_agent_id.is_not(None),
-        )
-    )
-
-
-def list_members(fleet_id: int) -> list[dict]:
-    """Return the fleet's active members, with placements.
-
-    A fleet has exactly one Director (the root), so every member placement's
-    ``director_agent_id`` equals the fleet root. Members are selected by
-    ``director_agent_id IS NOT NULL``, which lists all members and excludes
-    the root Director's own placement (the only row with a ``NULL`` director).
-
-    Args:
-        fleet_id: Fleet id to scope the query to.
-
-    Returns:
-        List of dicts each carrying ``agent_id``, ``name``, ``description``,
-        ``status``, ``registered_at``, and ``placement``.
-    """
-    stmt = _base_members_select(fleet_id)
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-    return [
-        {
-            "agent_id": row.agent_id,
-            "name": row.name,
-            "description": row.description,
-            "status": row.status,
-            "registered_at": row.registered_at,
-            "placement": _placement_dict(row),
-        }
-        for row in rows
-    ]
-
-
-def list_members_with_activity(fleet_id: int) -> list[dict]:
-    """``list_members`` plus per-member activity proxies sourced from ``tasks``.
-
-    Scoped by ``fleet_id`` only — the flat single-Director model means every
-    member placement's ``director_agent_id`` equals the fleet root, so the
-    ``director_agent_id IS NOT NULL`` member filter applies here too (and
-    excludes the root Director's own placement).
-
-    ``last_sent`` / ``last_recv`` / ``last_ack`` aggregate ``status_timestamp``
-    over the ``tasks`` table per agent. All three filter ``Task.type !=
-    'broadcast_summary'`` (mirrors ``poll_tasks``); broadcast_summary rows
-    land in the broadcaster's own context with ``status_state='completed'``
-    and would otherwise pollute every proxy for the broadcaster.
-
-    Args:
-        fleet_id: Fleet id to scope the query to.
-
-    Returns:
-        List of dicts as in :func:`list_members`, additionally carrying
-        ``last_sent``, ``last_recv``, ``last_ack`` (ISO timestamps or
-        ``None``), and ``idle`` — the integer-second delta between ``now``
-        and the most recent of ``last_sent`` / ``last_recv``, or ``None``
-        when both are ``None``.
-    """
-    last_sent_sq = (
-        select(func.max(Task.status_timestamp))
-        .where(
-            Task.from_agent_id == Agent.agent_id,
-            _NOT_BROADCAST_SUMMARY,
-        )
-        .correlate(Agent)
-        .scalar_subquery()
-    )
-    last_recv_sq = (
-        select(func.max(Task.status_timestamp))
-        .where(
-            Task.context_id == Agent.agent_id,
-            _NOT_BROADCAST_SUMMARY,
-        )
-        .correlate(Agent)
-        .scalar_subquery()
-    )
-    last_ack_sq = (
-        select(func.max(Task.status_timestamp))
-        .where(
-            Task.context_id == Agent.agent_id,
-            Task.status_state == "completed",
-            _NOT_BROADCAST_SUMMARY,
-        )
-        .correlate(Agent)
-        .scalar_subquery()
-    )
-    stmt = _base_members_select(fleet_id).add_columns(
-        last_sent_sq.label("last_sent"),
-        last_recv_sq.label("last_recv"),
-        last_ack_sq.label("last_ack"),
-    )
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-
-    now = datetime.now(UTC)
-    return [
-        {
-            "agent_id": row.agent_id,
-            "name": row.name,
-            "description": row.description,
-            "status": row.status,
-            "registered_at": row.registered_at,
-            "placement": _placement_dict(row),
-            "last_sent": row.last_sent,
-            "last_recv": row.last_recv,
-            "last_ack": row.last_ack,
-            "idle": _idle_seconds(now, row.last_sent, row.last_recv),
-        }
-        for row in rows
-    ]
-
-
-def _idle_seconds(
-    now: datetime, last_sent: str | None, last_recv: str | None
-) -> int | None:
-    candidates = [t for t in (last_sent, last_recv) if t is not None]
-    if not candidates:
-        return None
-    most_recent = datetime.fromisoformat(max(candidates))
-    delta = (now - most_recent).total_seconds()
-    return max(0, int(delta))
-
-
-def verify_agent_fleet(agent_id: int, fleet_id: int) -> bool:
-    """Return True iff the agent belongs to the fleet (any status).
-
-    Args:
-        agent_id: Agent id to verify.
-        fleet_id: Fleet id to check membership against.
-
-    Returns:
-        ``True`` if a matching row exists; ``False`` otherwise. Status is
-        ignored — deregistered agents still pass.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        return session.execute(
-            select(
-                exists().where(
-                    Agent.agent_id == agent_id,
-                    Agent.fleet_id == fleet_id,
-                )
-            )
-        ).scalar_one()
-
-
-def _row_to_task_dict(row) -> dict:
-    return {col: getattr(row, col) for col in _TASK_COLUMNS}
-
-
-def _insert_task(session, task_dict: dict) -> int:
-    """INSERT a new task without a ``task_id`` and return the DB-assigned id."""
-    return session.execute(
-        sqlite_insert(Task)
-        .values(
-            context_id=task_dict["context_id"],
-            from_agent_id=task_dict["from_agent_id"],
-            to_agent_id=task_dict["to_agent_id"],
-            type=task_dict["type"],
-            created_at=task_dict["created_at"],
-            status_state=task_dict["status_state"],
-            status_timestamp=task_dict["status_timestamp"],
-            origin_task_id=task_dict["origin_task_id"],
-            text=task_dict["text"],
-        )
-        .returning(Task.task_id)
-    ).scalar_one()
-
-
-def _save_task(session, task_dict: dict) -> None:
-    """UPDATE an existing task's mutable fields, keyed by ``task_id``."""
-    session.execute(
-        update(Task)
-        .where(Task.task_id == task_dict["task_id"])
-        .values(
-            status_state=task_dict["status_state"],
-            status_timestamp=task_dict["status_timestamp"],
-            origin_task_id=task_dict["origin_task_id"],
-            text=task_dict["text"],
-        )
-    )
-
-
-def _read_task(session, task_id: int) -> dict | None:
-    row = session.execute(
-        select(*(getattr(Task, col) for col in _TASK_COLUMNS)).where(
-            Task.task_id == task_id
-        )
-    ).first()
-    if row is None:
-        return None
-    return _row_to_task_dict(row)
-
-
-def _unicast_task_dict(
-    *,
-    recipient_id: int,
-    sender_id: int,
-    text: str,
-    now: str,
-    origin_task_id: int | None = None,
-) -> dict:
-    return {
-        "context_id": recipient_id,
-        "from_agent_id": sender_id,
-        "to_agent_id": recipient_id,
-        "type": "unicast",
-        "created_at": now,
-        "status_state": "input_required",
-        "status_timestamp": now,
-        "origin_task_id": origin_task_id,
-        "text": text,
-    }
-
-
-def send_message(fleet_id: int, agent_id: int, to: int | str, text: str) -> dict:
-    """Create a unicast task addressed to ``to`` and best-effort notify it.
-
-    Persists a new ``Task`` row with ``type='unicast'`` and
-    ``status_state='input_required'``, then calls
-    ``_try_notify_recipient`` to keystroke an inline preview into the
-    recipient's tmux pane. Notification failure does not roll back the
-    insert — the message remains available via :func:`poll_tasks`.
-
-    Args:
-        fleet_id: Fleet id; sender and recipient must both belong to it.
-        agent_id: Sender's agent id.
-        to: Recipient's agent id. Accepts a string for non-CLI callers
-            (WebUI, tests); it is coerced with ``int(...)``.
-        text: Message body. Truncation is render-side; the persisted row
-            holds the full string.
-
-    Returns:
-        Dict with ``task`` (the persisted task dict) and ``notification_sent``
-        (boolean indicating whether the inline-preview keystroke landed).
-
-    Raises:
-        ValueError: If ``to`` is not a valid integer, the sender is not
-            active in ``fleet_id``, or the recipient is missing or lives in a
-            different fleet.
-    """
-    try:
-        to_id = int(to)
-    except (TypeError, ValueError) as exc:
-        raise ValueError(f"Invalid destination format: {to}") from exc
-
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        if not _agent_is_active_in_fleet(session, agent_id, fleet_id):
-            raise ValueError(
-                f"Sender agent not found or not active in fleet: {agent_id}"
-            )
-
-        dest_fleet = session.execute(
-            select(Agent.fleet_id).where(
-                Agent.agent_id == to_id,
-                Agent.status == "active",
-            )
-        ).scalar_one_or_none()
-        if dest_fleet is None:
-            raise ValueError(f"Destination agent not found: {to_id}")
-        if dest_fleet != fleet_id:
-            raise ValueError(f"Destination agent not in fleet: {to_id}")
-
-        task_dict = _unicast_task_dict(
-            recipient_id=to_id,
-            sender_id=agent_id,
-            text=text,
-            now=_now_iso(),
-        )
-        task_dict["task_id"] = _insert_task(session, task_dict)
-        notification_sent = _try_notify_recipient(
-            session,
-            recipient_id=to_id,
-            sender_id=agent_id,
-            task_dict=task_dict,
-        )
-
-    return {"task": task_dict, "notification_sent": notification_sent}
-
-
-def broadcast_message(fleet_id: int, agent_id: int, text: str) -> list[dict]:
-    """Fan out one delivery task per active non-admin peer plus a sender summary.
-
-    Administrators are excluded at the SQL layer via ``json_extract`` so the
-    card blob stays in the database; they are write-only identities. Every
-    delivery row shares the same ``origin_task_id`` (the summary's task id)
-    so receivers can thread back to the original broadcast.
-
-    Args:
-        fleet_id: Fleet id to scope the broadcast to.
-        agent_id: Broadcaster's agent id.
-        text: Message body delivered to every recipient.
-
-    Returns:
-        Single-element list containing a dict with ``task`` (the summary row
-        owned by the broadcaster) and ``notifications_sent_count`` — the
-        number of inline-preview keystrokes that landed successfully.
-
-    Raises:
-        ValueError: If the sender is not active in ``fleet_id``.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        if not _agent_is_active_in_fleet(session, agent_id, fleet_id):
-            raise ValueError(
-                f"Sender agent not found or not active in fleet: {agent_id}"
-            )
-
-        recipient_ids = list(
-            session.execute(
-                select(Agent.agent_id).where(
-                    Agent.fleet_id == fleet_id,
-                    Agent.status == "active",
-                    Agent.agent_id != agent_id,
-                    func.coalesce(
-                        func.json_extract(Agent.agent_card_json, "$.cafleet.kind"),
-                        "",
-                    )
-                    != ADMINISTRATOR_KIND,
-                )
-            ).scalars()
-        )
-
-        now = _now_iso()
-        summary_dict = {
-            "context_id": agent_id,
-            "from_agent_id": agent_id,
-            "to_agent_id": 0,
-            "type": "broadcast_summary",
-            "created_at": now,
-            "status_state": "completed",
-            "status_timestamp": now,
-            "origin_task_id": None,
-            "text": f"Broadcast sent to {len(recipient_ids)} recipients",
-        }
-        summary_task_id = _insert_task(session, summary_dict)
-        summary_dict["task_id"] = summary_task_id
-        summary_dict["origin_task_id"] = summary_task_id
-        _save_task(session, summary_dict)
-
-        deliveries: list[tuple[int, dict]] = []
-        for recipient_id in recipient_ids:
-            delivery_dict = _unicast_task_dict(
-                recipient_id=recipient_id,
-                sender_id=agent_id,
-                text=text,
-                now=_now_iso(),
-                origin_task_id=summary_task_id,
-            )
-            delivery_dict["task_id"] = _insert_task(session, delivery_dict)
-            deliveries.append((recipient_id, delivery_dict))
-
-        notifications_sent_count = sum(
-            _try_notify_recipient(
-                session,
-                recipient_id=recipient_id,
-                sender_id=agent_id,
-                task_dict=delivery_dict,
-            )
-            for recipient_id, delivery_dict in deliveries
-        )
-
-    return [
-        {"task": summary_dict, "notifications_sent_count": notifications_sent_count}
-    ]
-
-
-def poll_tasks(agent_id: int) -> list[dict]:
-    """Return un-acked deliveries addressed to ``agent_id``, newest first.
-
-    Only ``input_required`` tasks are returned — once a delivery is ACKed
-    (``completed``) or canceled it no longer appears. ``broadcast_summary``
-    rows are filtered out, as those belong to the broadcaster's own context
-    and are not deliveries.
-
-    Args:
-        agent_id: Recipient agent id; matches ``Task.context_id``.
-
-    Returns:
-        List of flat task dicts (one per row) carrying every column from the
-        ``tasks`` table, in DESC ``status_timestamp`` order.
-    """
-    return _list_tasks_where(
-        Task.context_id == agent_id,
-        status="input_required",
-    )
-
-
-def _transition_task_state(
-    agent_id: int,
-    task_id: int,
-    *,
-    expected_agent_field: str,
-    new_state: str,
-    action_verb: str,
-    permission_error_msg: str,
-) -> dict:
-    sm = get_sync_sessionmaker()
-    with sm() as session, session.begin():
-        task_dict = _read_task(session, task_id)
-        if task_dict is None:
-            raise ValueError(f"Task {task_id} not found")
-
-        if task_dict[expected_agent_field] != agent_id:
-            raise PermissionError(permission_error_msg)
-
-        if task_dict["status_state"] != "input_required":
-            raise ValueError(
-                f"Cannot {action_verb} task in state {task_dict['status_state']}"
-            )
-
-        task_dict["status_state"] = new_state
-        task_dict["status_timestamp"] = _now_iso()
-
-        _save_task(session, task_dict)
-
-    return {"task": task_dict}
-
-
-def ack_task(agent_id: int, task_id: int) -> dict:
-    """Transition a task from ``input_required`` to ``completed`` for the recipient.
-
-    Args:
-        agent_id: Recipient agent id; must match ``Task.context_id``.
-        task_id: Task id to ack.
-
-    Returns:
-        Dict with ``task`` — the updated task dict.
-
-    Raises:
-        ValueError: If the task does not exist or is not in
-            ``input_required`` state.
-        PermissionError: If ``agent_id`` is not the recipient.
-    """
-    return _transition_task_state(
-        agent_id,
-        task_id,
-        expected_agent_field="context_id",
-        new_state="completed",
-        action_verb="ACK",
-        permission_error_msg="Only the recipient can ACK a task",
-    )
-
-
-def cancel_task(agent_id: int, task_id: int) -> dict:
-    """Transition a task from ``input_required`` to ``canceled`` for the sender.
-
-    Args:
-        agent_id: Sender agent id; must match ``Task.from_agent_id``.
-        task_id: Task id to cancel.
-
-    Returns:
-        Dict with ``task`` — the updated task dict.
-
-    Raises:
-        ValueError: If the task does not exist or is not in
-            ``input_required`` state.
-        PermissionError: If ``agent_id`` is not the sender.
-    """
-    return _transition_task_state(
-        agent_id,
-        task_id,
-        expected_agent_field="from_agent_id",
-        new_state="canceled",
-        action_verb="cancel",
-        permission_error_msg="Only the sender can cancel a task",
-    )
-
-
-def list_fleet_agents(fleet_id: int) -> list[dict]:
-    """Return active agents plus deregistered agents that still own tasks.
-
-    ``kind`` is derived in SQL via ``json_extract`` so the card blob never
-    leaves SQLite — otherwise we would materialize every row's JSON just to
-    compute a one-token discriminator. ``coalesce`` handles cards without a
-    ``cafleet.kind`` path by substituting an empty string.
-    """
-    has_tasks = exists().where(
-        or_(
-            Task.context_id == Agent.agent_id,
-            Task.from_agent_id == Agent.agent_id,
-        )
-    )
-    kind_expr = func.coalesce(
-        func.json_extract(Agent.agent_card_json, "$.cafleet.kind"), ""
-    )
-    stmt = select(
-        Agent.agent_id,
-        Agent.name,
-        Agent.description,
-        Agent.status,
-        Agent.registered_at,
-        kind_expr.label("kind_raw"),
-    ).where(
-        Agent.fleet_id == fleet_id,
-        or_(
-            Agent.status == "active",
-            and_(Agent.status == "deregistered", has_tasks),
-        ),
-    )
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-    return [
-        {
-            "agent_id": row.agent_id,
-            "name": row.name,
-            "description": row.description,
-            "status": row.status,
-            "registered_at": row.registered_at,
-            "kind": (
-                ADMINISTRATOR_KIND if row.kind_raw == ADMINISTRATOR_KIND else "user"
-            ),
-        }
-        for row in rows
-    ]
-
-
-def _list_tasks_where(
-    *filters,
-    status: str | None = None,
-) -> list[dict]:
-    stmt = (
-        select(*(getattr(Task, col) for col in _TASK_COLUMNS))
-        .where(*filters, _NOT_BROADCAST_SUMMARY)
-        .order_by(Task.status_timestamp.desc())
-    )
-    if status is not None:
-        stmt = stmt.where(Task.status_state == status)
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-    return [_row_to_task_dict(row) for row in rows]
-
-
-def list_inbox(agent_id: int) -> list[dict]:
-    """Return raw task rows addressed to ``agent_id`` (no broadcast_summary)."""
-    return _list_tasks_where(Task.context_id == agent_id)
-
-
-def list_sent(agent_id: int) -> list[dict]:
-    """Return raw task rows sent by ``agent_id`` (no broadcast_summary)."""
-    return _list_tasks_where(Task.from_agent_id == agent_id)
-
-
-def list_timeline(fleet_id: int, limit: int = 200) -> list[dict]:
-    """Return the fleet's recent tasks in DESC ``status_timestamp`` order.
-
-    ``broadcast_summary`` rows are filtered out so the timeline shows only
-    delivery rows. Membership is tested via ``from_agent_id`` joined to
-    ``agents.fleet_id``.
-
-    Args:
-        fleet_id: Fleet id to scope the query to.
-        limit: Maximum number of rows to return (default 200).
-
-    Returns:
-        List of flat task dicts in DESC ``status_timestamp`` order.
-    """
-    stmt = (
-        select(*(getattr(Task, col) for col in _TASK_COLUMNS))
-        .join(Agent, Task.from_agent_id == Agent.agent_id)
-        .where(
-            Agent.fleet_id == fleet_id,
-            _NOT_BROADCAST_SUMMARY,
-        )
-        .order_by(Task.status_timestamp.desc())
-        .limit(limit)
-    )
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(stmt).all()
-    return [_row_to_task_dict(row) for row in rows]
-
-
-def get_agent_names(agent_ids: list[int]) -> dict[int, str]:
-    """Batch ``agent_id → name`` lookup including deregistered agents."""
-    if not agent_ids:
-        return {}
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        rows = session.execute(
-            select(Agent.agent_id, Agent.name).where(Agent.agent_id.in_(agent_ids))
-        ).all()
-    return {row.agent_id: row.name for row in rows}
-
-
-def get_task(fleet_id: int, task_id: int) -> dict:
-    """Return the task iff at least one of its endpoints lives in the fleet.
-
-    Args:
-        fleet_id: Fleet id used to gate visibility.
-        task_id: Task id to fetch.
-
-    Returns:
-        Dict with ``task`` — the flat typed-column task dict.
-
-    Raises:
-        ValueError: If the task does not exist or neither endpoint belongs
-            to ``fleet_id``.
-    """
-    sm = get_sync_sessionmaker()
-    with sm() as session:
-        task_dict = _read_task(session, task_id)
-        if task_dict is None:
-            raise ValueError(f"Task {task_id} not found")
-
-        endpoint_ids = [task_dict["from_agent_id"]]
-        to_id = task_dict["to_agent_id"]
-        if to_id:
-            endpoint_ids.append(to_id)
-        in_fleet = session.execute(
-            select(
-                exists().where(
-                    Agent.agent_id.in_(endpoint_ids),
-                    Agent.fleet_id == fleet_id,
-                )
-            )
-        ).scalar_one()
-        if not in_fleet:
-            raise ValueError(f"Task {task_id} not found")
-
-    return {"task": task_dict}
